Move Bias link from Caffe function helpers to standard links.
